// src/app/models/interface/personal-contratista-request.ts
export interface PersonalContratistaRequest {

  numeroDocumento?: string;
  codigoEmpleado?: number;
  nombres?: string;
  apellidoPaterno?: string;
  apellidoMaterno?: string;
  idEmpresa?: number;
  codigoCargo?: string;
  codigoOficina?: string;
  periodoIngreso?: string;
  estadoLaboral?: string;
  estadoSedapal?: string;
  codMotivoCese?: number;
  fechaCese?: string;
  tipoSolicitud?: number;
  estadoSolicitud?: string;
  idPersonal?: number;
  idPerfil?: number;
  codLote?: string;

}